Fixes issue 91 by omitting the head element's start tag when the head element is empty.
Also adds checks againts "EmptyTag" when the expected "start tag" might be the one of a "void element" (this is an undocumented mess in html5lib between what should be a StartTag vs. an EmptyTag and a Characters vs. SpaceCharacters token; the "optionaltags" filter expects what is emitted by a treewalker).

Therefore, fixed a few tests that used "EmptyTag" for to represent empty (no child) non-void elements.

git-svn-id: http://html5lib.googlecode.com/svn/trunk@1271 acbfec75-9323-0410-a652-858a13e371e0